+++ e2e/cypress/dataHooks.js
@@ -5,6 +5,8 @@ export const INLINE_TOOLBAR_BUTTONS = {
   COLOR: 'TextColorButton',
   TITLE: 'textBlockStyleButton_Title',
   QUOTE: 'textBlockStyleButton_Quote',
+  DECREASE_INDENT: 'decreaseIndentButton',
+  INCREASE_INDENT: 'increaseIndentButton',
   ORDERED_LIST: 'textBlockStyleButton_Numberedlist',
   UNORDERED_LIST: 'textBlockStyleButton_Bulletedlist',
   ALIGNMENT: 'textDropDownButton_Alignment',