+++ packages/editor-common/web/src/lib/UrlInputModal.js
@@ -17,6 +17,7 @@ export default class UrlInputModal extends Component {
   onUrlChange = event => {
     const url = event.target.value;
     this.props.onInputChange(url);
+    this.setState({ isDropdownOpen: url !== '' });
   };
 
   handleKeyPress = event => {