+++ packages/editor-common/web/src/index.js
@@ -95,6 +95,8 @@ export {
   getPostContentSummary,
   createSelection,
   getBlockType,
+  indentSelectedBlocks,
+  isTypeText,
   setForceSelection,
   deleteBlockText,
   insertString,