+++ packages/editor/web/src/RichContentEditor/Toolbars/buttons/index.js
@@ -26,7 +26,6 @@ export {
   boldButton,
   italicButton,
   underlineButton,
-  indentButton,
   titleButton,
   blockquoteButton,
   alignTextLeftButton,