+++ packages/plugin-indent/web/src/defaults.js
@@ -1,0 +1,9 @@
+import { INDENT_TYPE as type } from './types';
+export const DEFAULTS = Object.freeze({
+  type,
+  config: {},
+});
+
+export const THEME = () => {
+  return {};
+};