+++ CHANGELOG.md
@@ -16,8 +16,14 @@
   </summary>
 
 ### :rocket: New Feature
+- `indent`
+  - [#898](https://github.com/wix-incubator/rich-content/pull/898) text indentation
 - `viewer`
   - [#1005](https://github.com/wix-incubator/rich-content/pull/1005) add viewMode SEO to ProGallery
+
+### :bug: Bug Fix
+- `gallery`
+  - [#1020](https://github.com/wix-incubator/rich-content/pull/1020) gallery accepts window as scrollingElement
 
 </details>
 <hr/>