+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 ## 7.6.0 (May 5, 2020)
 ### :rocket: New Feature
 - `indent`
-  - [#898](https://github.com/wix-incubator/rich-content/pull/898) alpha text indentation
+  - [#898](https://github.com/wix-incubator/rich-content/pull/898) text indentation
 - `viewer`
   - [#1005](https://github.com/wix-incubator/rich-content/pull/1005) add viewMode SEO to ProGallery
 ### :bug: Bug Fix