+++ CHANGELOG.md
@@ -14,10 +14,26 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
-
-
+  
+### :bug: Bug Fix
+- `editor-common`
+  - [#1046](https://github.com/wix-incubator/rich-content/pull/1046) fix: cursor jumps to start of editor on close modal
 </details>
 <hr/>
+
+## 7.6.0 (May 5, 2020)
+### :rocket: New Feature
+- `indent`
+  - [#898](https://github.com/wix-incubator/rich-content/pull/898) text indentation
+- `viewer`
+  - [#1005](https://github.com/wix-incubator/rich-content/pull/1005) add viewMode SEO to ProGallery
+### :bug: Bug Fix
+- `social-modals`
+  - [#1037](https://github.com/wix-incubator/rich-content/pull/1037) disable text input autocomplete
+- `vertical embed`
+  - [#1036](https://github.com/wix-incubator/rich-content/pull/1036) dropdown for search opens automatically when typing text for search
+- `gallery`
+  - [#1020](https://github.com/wix-incubator/rich-content/pull/1020) gallery accepts window as scrollingElement
 
 ## 7.5.0 (May 5, 2020)
 ### :rocket: New Feature
@@ -32,6 +48,8 @@
   - [#941](https://github.com/wix-incubator/rich-content/pull/941) add tooltips for settings panels
 - `preview`
   - [#999](https://github.com/wix-incubator/rich-content/pull/999) giphy metadata is handled correctly by image and gallery data mergers
+- `gallery`
+  - [#1006](https://github.com/wix-incubator/rich-content/pull/1006) expand icon appears only on hovered image
 ### :house: Internal
 - `wrapper`
   - [#980](https://github.com/wix-incubator/rich-content/pull/980) createEmpty import